+++ lib/screens/authentication/email_password/email_password.dart
@@ -1,0 +1,3 @@
+export 'register_screen.dart';
+export 'sign_in_screen.dart';
+export 'user_info_screen.dart';
