+++ lib/screens/authentication/google_sign_in/google_sign_in.dart
@@ -1,0 +1,2 @@
+export 'sign_in_screen.dart';
+export 'user_info_screen.dart';
