+++ lib/widgets/authentication/email_password/register_form.dart
@@ -1,48 +1,83 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutterfire_samples/res/custom_colors.dart';
-import 'package:flutterfire_samples/screens/authentication/email_password/ep_register_screen.dart';
-import 'package:flutterfire_samples/screens/authentication/email_password/ep_user_info_screen.dart';
 import 'package:flutterfire_samples/utils/ep_authentication.dart';
 import 'package:flutterfire_samples/utils/ep_validator.dart';
 
+import '../../../screens/authentication/email_password/email_password.dart';
 import '../../custom_form_field.dart';
 
-class EPSignInForm extends StatefulWidget {
+class RegisterForm extends StatefulWidget {
+  final FocusNode nameFocusNode;
   final FocusNode emailFocusNode;
   final FocusNode passwordFocusNode;
 
-  const EPSignInForm({
+  const RegisterForm({
     Key? key,
+    required this.nameFocusNode,
     required this.emailFocusNode,
     required this.passwordFocusNode,
   }) : super(key: key);
   @override
-  _EPSignInFormState createState() => _EPSignInFormState();
+  _RegisterFormState createState() => _RegisterFormState();
 }
 
-class _EPSignInFormState extends State<EPSignInForm> {
+class _RegisterFormState extends State<RegisterForm> {
+  final TextEditingController _nameController = TextEditingController();
   final TextEditingController _emailController = TextEditingController();
   final TextEditingController _passwordController = TextEditingController();
 
-  final _signInFormKey = GlobalKey<FormState>();
+  final _registerFormKey = GlobalKey<FormState>();
 
-  bool _isSigningIn = false;
+  bool _isSigningUp = false;
+
+  Route _routeToSignInScreen() {
+    return PageRouteBuilder(
+      pageBuilder: (context, animation, secondaryAnimation) =>
+          const SignInScreen(),
+      transitionsBuilder: (context, animation, secondaryAnimation, child) {
+        var begin = const Offset(-1.0, 0.0);
+        var end = Offset.zero;
+        var curve = Curves.ease;
+
+        var tween =
+            Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
+
+        return SlideTransition(
+          position: animation.drive(tween),
+          child: child,
+        );
+      },
+    );
+  }
 
   @override
   Widget build(BuildContext context) {
     return Form(
-      key: _signInFormKey,
+      key: _registerFormKey,
       child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
         children: [
           Padding(
             padding: const EdgeInsets.only(
               left: 8.0,
               right: 8.0,
-              bottom: 24.0,
             ),
             child: Column(
               children: [
+                CustomFormField(
+                  controller: _nameController,
+                  focusNode: widget.nameFocusNode,
+                  keyboardType: TextInputType.name,
+                  inputAction: TextInputAction.next,
+                  isCapitalized: true,
+                  validator: (value) => EPValidator.validateName(
+                    name: value,
+                  ),
+                  label: 'Name',
+                  hint: 'Enter your name',
+                ),
+                const SizedBox(height: 16.0),
                 CustomFormField(
                   controller: _emailController,
                   focusNode: widget.emailFocusNode,
@@ -70,7 +105,8 @@ class _EPSignInFormState extends State<EPSignInForm> {
               ],
             ),
           ),
-          _isSigningIn
+          const SizedBox(height: 24.0),
+          _isSigningUp
               ? Padding(
                   padding: const EdgeInsets.all(16.0),
                   child: CircularProgressIndicator(
@@ -99,21 +135,22 @@ class _EPSignInFormState extends State<EPSignInForm> {
                         widget.passwordFocusNode.unfocus();
 
                         setState(() {
-                          _isSigningIn = true;
+                          _isSigningUp = true;
                         });
 
-                        if (_signInFormKey.currentState!.validate()) {
+                        if (_registerFormKey.currentState!.validate()) {
                           User? user =
-                              await EPAuthentication.signInUsingEmailPassword(
-                            context: context,
+                              await EPAuthentication.registerUsingEmailPassword(
+                            name: _nameController.text,
                             email: _emailController.text,
                             password: _passwordController.text,
+                            context: context,
                           );
 
                           if (user != null) {
                             Navigator.of(context).pushReplacement(
                               MaterialPageRoute(
-                                builder: (context) => EPUserInfoScreen(
+                                builder: (context) => UserInfoScreen(
                                   user: user,
                                 ),
                               ),
@@ -122,13 +159,13 @@ class _EPSignInFormState extends State<EPSignInForm> {
                         }
 
                         setState(() {
-                          _isSigningIn = false;
+                          _isSigningUp = false;
                         });
                       },
                       child: Padding(
                         padding: const EdgeInsets.only(top: 16.0, bottom: 16.0),
                         child: Text(
-                          'LOGIN',
+                          'REGISTER',
                           style: TextStyle(
                             fontSize: 24,
                             fontWeight: FontWeight.bold,
@@ -144,13 +181,11 @@ class _EPSignInFormState extends State<EPSignInForm> {
           InkWell(
             onTap: () {
               Navigator.of(context).pushReplacement(
-                MaterialPageRoute(
-                  builder: (context) => EPRegisterScreen(),
-                ),
+                _routeToSignInScreen(),
               );
             },
             child: Text(
-              'Don\'t have an account? Sign up',
+              'Already have an account? Sign in',
               style: TextStyle(
                 color: Palette.firebaseGrey,
                 letterSpacing: 0.5,
