+++ lib/widgets/authentication/email_password/sign_in_form.dart
@@ -1,83 +1,47 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutterfire_samples/res/custom_colors.dart';
-import 'package:flutterfire_samples/screens/authentication/email_password/ep_sign_in_screen.dart';
-import 'package:flutterfire_samples/screens/authentication/email_password/ep_user_info_screen.dart';
 import 'package:flutterfire_samples/utils/ep_authentication.dart';
 import 'package:flutterfire_samples/utils/ep_validator.dart';
 
+import '../../../screens/authentication/email_password/email_password.dart';
 import '../../custom_form_field.dart';
 
-class EPRegisterForm extends StatefulWidget {
-  final FocusNode nameFocusNode;
+class SignInForm extends StatefulWidget {
   final FocusNode emailFocusNode;
   final FocusNode passwordFocusNode;
 
-  const EPRegisterForm({
+  const SignInForm({
     Key? key,
-    required this.nameFocusNode,
     required this.emailFocusNode,
     required this.passwordFocusNode,
   }) : super(key: key);
   @override
-  _EPRegisterFormState createState() => _EPRegisterFormState();
+  _SignInFormState createState() => _SignInFormState();
 }
 
-class _EPRegisterFormState extends State<EPRegisterForm> {
-  final TextEditingController _nameController = TextEditingController();
+class _SignInFormState extends State<SignInForm> {
   final TextEditingController _emailController = TextEditingController();
   final TextEditingController _passwordController = TextEditingController();
 
-  final _registerFormKey = GlobalKey<FormState>();
+  final _signInFormKey = GlobalKey<FormState>();
 
-  bool _isSigningUp = false;
-
-  Route _routeToSignInScreen() {
-    return PageRouteBuilder(
-      pageBuilder: (context, animation, secondaryAnimation) => EPSignInScreen(),
-      transitionsBuilder: (context, animation, secondaryAnimation, child) {
-        var begin = const Offset(-1.0, 0.0);
-        var end = Offset.zero;
-        var curve = Curves.ease;
-
-        var tween =
-            Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
-
-        return SlideTransition(
-          position: animation.drive(tween),
-          child: child,
-        );
-      },
-    );
-  }
+  bool _isSigningIn = false;
 
   @override
   Widget build(BuildContext context) {
     return Form(
-      key: _registerFormKey,
+      key: _signInFormKey,
       child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
         children: [
           Padding(
             padding: const EdgeInsets.only(
               left: 8.0,
               right: 8.0,
+              bottom: 24.0,
             ),
             child: Column(
               children: [
-                CustomFormField(
-                  controller: _nameController,
-                  focusNode: widget.nameFocusNode,
-                  keyboardType: TextInputType.name,
-                  inputAction: TextInputAction.next,
-                  isCapitalized: true,
-                  validator: (value) => EPValidator.validateName(
-                    name: value,
-                  ),
-                  label: 'Name',
-                  hint: 'Enter your name',
-                ),
-                const SizedBox(height: 16.0),
                 CustomFormField(
                   controller: _emailController,
                   focusNode: widget.emailFocusNode,
@@ -105,10 +69,9 @@ class _EPRegisterFormState extends State<EPRegisterForm> {
               ],
             ),
           ),
-          const SizedBox(height: 24.0),
-          _isSigningUp
-              ? Padding(
-                  padding: const EdgeInsets.all(16.0),
+          _isSigningIn
+              ? const Padding(
+                  padding: EdgeInsets.all(16.0),
                   child: CircularProgressIndicator(
                     valueColor: AlwaysStoppedAnimation<Color>(
                       Palette.firebaseOrange,
@@ -116,7 +79,7 @@ class _EPRegisterFormState extends State<EPRegisterForm> {
                   ),
                 )
               : Padding(
-                  padding:const  EdgeInsets.only(left: 0.0, right: 0.0),
+                  padding: const EdgeInsets.only(left: 0.0, right: 0.0),
                   child: SizedBox(
                     width: double.maxFinite,
                     child: ElevatedButton(
@@ -135,22 +98,21 @@ class _EPRegisterFormState extends State<EPRegisterForm> {
                         widget.passwordFocusNode.unfocus();
 
                         setState(() {
-                          _isSigningUp = true;
+                          _isSigningIn = true;
                         });
 
-                        if (_registerFormKey.currentState!.validate()) {
+                        if (_signInFormKey.currentState!.validate()) {
                           User? user =
-                              await EPAuthentication.registerUsingEmailPassword(
-                            name: _nameController.text,
+                              await EPAuthentication.signInUsingEmailPassword(
+                            context: context,
                             email: _emailController.text,
                             password: _passwordController.text,
-                            context: context,
                           );
 
                           if (user != null) {
                             Navigator.of(context).pushReplacement(
                               MaterialPageRoute(
-                                builder: (context) => EPUserInfoScreen(
+                                builder: (context) => UserInfoScreen(
                                   user: user,
                                 ),
                               ),
@@ -159,13 +121,13 @@ class _EPRegisterFormState extends State<EPRegisterForm> {
                         }
 
                         setState(() {
-                          _isSigningUp = false;
+                          _isSigningIn = false;
                         });
                       },
-                      child: Padding(
-                        padding: const EdgeInsets.only(top: 16.0, bottom: 16.0),
+                      child: const Padding(
+                        padding: EdgeInsets.only(top: 16.0, bottom: 16.0),
                         child: Text(
-                          'REGISTER',
+                          'LOGIN',
                           style: TextStyle(
                             fontSize: 24,
                             fontWeight: FontWeight.bold,
@@ -181,11 +143,13 @@ class _EPRegisterFormState extends State<EPRegisterForm> {
           InkWell(
             onTap: () {
               Navigator.of(context).pushReplacement(
-                _routeToSignInScreen(),
+                MaterialPageRoute(
+                  builder: (context) => const RegisterScreen(),
+                ),
               );
             },
-            child: Text(
-              'Already have an account? Sign in',
+            child: const Text(
+              'Don\'t have an account? Sign up',
               style: TextStyle(
                 color: Palette.firebaseGrey,
                 letterSpacing: 0.5,
